+++ lib/src/theme/themes/default_theme_no_secondary_border_variant.dart
@@ -504,6 +504,7 @@ class ShadDefaultThemeNoSecondaryBorderVariant extends ShadThemeVariant {
       size: 16,
       duration: 100.milliseconds,
       color: colorScheme.primary,
+      uncheckedColor: colorScheme.input,
       padding: const EdgeInsetsDirectional.only(start: 8),
       decoration: ShadDecoration(
         border: ShadBorder.all(