+++ lib/src/theme/themes/default_theme_variant.dart
@@ -488,6 +488,7 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
       size: 16,
       duration: 100.milliseconds,
       color: colorScheme.primary,
+      uncheckedColor: colorScheme.input,
       padding: const EdgeInsetsDirectional.only(start: 8),
       checkboxPadding: const EdgeInsets.only(top: 1),
       decoration: ShadDecoration(