+++ lib/src/utils/clipboard/clipboard_service_stub.dart
@@ -1,0 +1,6 @@
+import 'package:shadcn_ui/src/utils/clipboard/clipboard_service.dart';
+
+void addPasteFilesListener(PasteFilesCallback callback) {}
+void removePasteFilesListener(PasteFilesCallback callback) {}
+void addPasteFilesErrorListener(PasteFilesErrorCallback callback) {}
+void removePasteFilesErrorListener(PasteFilesErrorCallback callback) {}